test: make sure the sourcemap fetcher is available before running assertions
