+++ systemtest/sourcemap_test.go
@@ -21,7 +21,6 @@ import (
 	"os"
 	"testing"
 
-	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
 	"github.com/elastic/apm-server/systemtest"
@@ -40,7 +39,7 @@ func TestRUMErrorSourcemapping(t *testing.T) {
 
 		test := func(t *testing.T, serverURL string) {
 			systemtest.SendRUMEventsPayload(t, serverURL, "../testdata/intake-v2/errors_rum.ndjson")
-			result := systemtest.Elasticsearch.ExpectDocs(t, "logs-apm.error-*", nil)
+			result := systemtest.Elasticsearch.ExpectSourcemapError(t, "logs-apm.error-*", nil, true)
 			systemtest.ApproveEvents(
 				t, t.Name(), result.Hits.Hits,
 				// RUM timestamps are set by the server based on the time the payload is received.
@@ -86,10 +85,10 @@ func TestRUMSpanSourcemapping(t *testing.T) {
 	require.NoError(t, err)
 
 	systemtest.SendRUMEventsPayload(t, srv.URL, "../testdata/intake-v2/transactions_spans_rum_2.ndjson")
-	result := systemtest.Elasticsearch.ExpectDocs(t, "traces-apm*", estest.TermQuery{
+	result := systemtest.Elasticsearch.ExpectSourcemapError(t, "traces-apm*", estest.TermQuery{
 		Field: "processor.event",
 		Value: "span",
-	})
+	}, true)
 
 	systemtest.ApproveEvents(
 		t, t.Name(), result.Hits.Hits,
@@ -117,10 +116,10 @@ func TestNoMatchingSourcemap(t *testing.T) {
 	require.NoError(t, err)
 
 	systemtest.SendRUMEventsPayload(t, srv.URL, "../testdata/intake-v2/transactions_spans_rum_2.ndjson")
-	result := systemtest.Elasticsearch.ExpectDocs(t, "traces-apm*", estest.TermQuery{
+	result := systemtest.Elasticsearch.ExpectSourcemapError(t, "traces-apm*", estest.TermQuery{
 		Field: "processor.event",
 		Value: "span",
-	})
+	}, false)
 
 	systemtest.ApproveEvents(
 		t, t.Name(), result.Hits.Hits,
@@ -148,14 +147,12 @@ func TestSourcemapCaching(t *testing.T) {
 
 	// Index an error, applying source mapping and caching the source map in the process.
 	systemtest.SendRUMEventsPayload(t, srv.URL, "../testdata/intake-v2/errors_rum.ndjson")
-	result := systemtest.Elasticsearch.ExpectDocs(t, "logs-apm.error-*", nil)
-	assertSourcemapUpdated(t, result, true)
+	systemtest.Elasticsearch.ExpectSourcemapError(t, "logs-apm.error-*", nil, true)
 
 	// Delete the source map and error, and try again.
 	systemtest.CleanupElasticsearch(t)
 	systemtest.SendRUMEventsPayload(t, srv.URL, "../testdata/intake-v2/errors_rum.ndjson")
-	result = systemtest.Elasticsearch.ExpectMinDocs(t, 1, "logs-apm.error-*", nil)
-	assertSourcemapUpdated(t, result, true)
+	systemtest.Elasticsearch.ExpectSourcemapError(t, "logs-apm.error-*", nil, true)
 }
 
 func TestSourcemapFetcher(t *testing.T) {
@@ -207,8 +204,7 @@ func TestSourcemapFetcher(t *testing.T) {
 
 			// Index an error, applying source mapping and caching the source map in the process.
 			systemtest.SendRUMEventsPayload(t, srv.URL, "../testdata/intake-v2/errors_rum.ndjson")
-			result := systemtest.Elasticsearch.ExpectDocs(t, "logs-apm.error-*", nil)
-			assertSourcemapUpdated(t, result, true)
+			systemtest.Elasticsearch.ExpectSourcemapError(t, "logs-apm.error-*", nil, true)
 		})
 	}
 }
@@ -220,40 +216,4 @@ func deleteIndex(t *testing.T, name string) {
 	resp, err = systemtest.Elasticsearch.Indices.Flush()
 	require.NoError(t, err)
 	resp.Body.Close()
-}
-
-func assertSourcemapUpdated(t *testing.T, result estest.SearchResult, updated bool) {
-	t.Helper()
-
-	type StacktraceFrame struct {
-		Sourcemap struct {
-			Updated bool
-		}
-	}
-	type Error struct {
-		Exception []struct {
-			Stacktrace []StacktraceFrame
-		}
-		Log struct {
-			Stacktrace []StacktraceFrame
-		}
-	}
-
-	for _, hit := range result.Hits.Hits {
-		var source struct {
-			Error Error
-		}
-		err := hit.UnmarshalSource(&source)
-		require.NoError(t, err)
-
-		for _, exception := range source.Error.Exception {
-			for _, stacktrace := range exception.Stacktrace {
-				assert.Equal(t, updated, stacktrace.Sourcemap.Updated)
-			}
-		}
-
-		for _, stacktrace := range source.Error.Log.Stacktrace {
-			assert.Equal(t, updated, stacktrace.Sourcemap.Updated)
-		}
-	}
 }
